requirements.txt
django
psycopg2
django-crontab
pandas
python-dotenv
django-cors-headers
gunicorn
beautifulsoup4
requests
langchain
openai